Update FlairYourPostBot.py
Added in add_flair_message_wiki which looks for a wiki page in the subreddit that is titled "FlairYourPostBotInitial" and it uses that as the initial remove message. If the post is not flaired within 10 minutes than the bot uses remove_post_message_wiki which looks for a wiki page in the subreddit that is titles "FlairYourPostBotFinal" and it uses that as the final remove message. All subreddits will be required to make the two wiki pages and copy paste the template for the message, after that mods can choose to edit the template or leave it as it is.